MGS-4979 Make JSON schema validation work on CouchDB 2.x
The jsen library previously used for validation documents
against a JSON schema generated source for a validator,
then used `eval` to execute it. This doesn't work on
CouchDB 2.x, where CSP rules prevent this operation.

This change switches to the ajv library for JSON validation.
The ajv-pack library is used to generate the source for the
schema validator at build time, so does not rely on `eval`.

The generated validator requires some runtime classes from
ajv, and they are copied into the couch app directory by
the build.

This change also changes the program used to create the
CouchDB design document. This used to be couchdb-push,
which was installed as an NPM package. This has been
changed to couchapp, which can be installed using packages
included with the operating system.